==> notebooks/0.ipynb <==
# Tourism Data Processing
DS Senior Project - Noah B Johnson - Luther College

import re

import numpy as np
import pandas as pd
import pycountry

%matplotlib inline

pd.set_option('display.float_format', lambda x: '%.3f' % x)

## World Development Indicators

wdi = pd.read_excel(
    "../data/external/Tourism/WDI/Data_Extract_From_World_Development_Indicators.xlsx"
)

wdi.head(3)

wdi['Series Code'].unique()

wdi.tail(3)

# nan is not a valid series code, so we should remove those rows
wdi.dropna(subset=['Series Code', 'Country Code'], inplace=True)

wdi.tail(3)

### Pivoting the Data
> Target format is data by year-country with indicator columns

# Making our keys
country_codes = wdi['Country Code'].unique()
years = wdi.columns[4:].tolist()

print("We will have {} rows".format(len(country_codes) * len(years)))

""" Manual transposition of data
"""

year_column = []
country_column = []
data_columns = {
    'ST.INT.ARVL': [],
    'ST.INT.XPND.MP.ZS': [],
    'ST.INT.XPND.CD': [],
    'ST.INT.TRNX.CD': [],
    'ST.INT.TVLX.CD': [],
    'ST.INT.DPRT': [],
    'ST.INT.RCPT.XP.ZS': [],
    'ST.INT.RCPT.CD': [],
    'ST.INT.TRNR.CD': [],
    'ST.INT.TVLR.CD': []
}
for country in country_codes:
    for year in years:
        year_column.append(year)
        country_column.append(country)
        for indicator in data_columns.keys():
            cdf = wdi.loc[(wdi['Country Code'] == country)]
            datum = cdf[cdf['Series Code'] == indicator][year].tolist()[0]
            data_columns[indicator].append(datum)

""" Reassign country and year to columns
"""
data_columns["Year"] = year_column
data_columns["Country Code"] = country_column

""" Data is correctly formatted
"""
wdi_clean = pd.DataFrame(data_columns)

### Cleaning data
> fix nans, assign correct data types, and standardize countries

wdi_clean.head(3)

""" Replace '..' with np.nan for better parsing
"""
wdi_clean = wdi_clean.replace('..', np.NaN)

wdi_clean.head(3)

""" Assign all columns except year and country to float dtype
"""
wdi_clean[wdi_clean.columns[:-2]] = wdi_clean[wdi_clean.columns[:-2]].astype(
    float)

""" Check the dtypes of the dataframe
"""
wdi_clean.get_dtype_counts()

""" Remove the bracketed text and parse year into an integer so: '1960 [YR1960]' → int(1960)
"""
wdi_clean["Year"] = wdi_clean["Year"].apply(lambda x: re.match(r'^\d*', x)[0]
                                            ).astype(int)

""" Check year values
"""
wdi_clean.sample(3)

""" 
Create a mask without regions and custom groupings, 
selecting only valid country entries
"""
country_bool = []
for country in wdi_clean['Country Code'].tolist():
    try:
        pycountry.countries.lookup(country)
        country_bool.append(True)
    except LookupError:
        country_bool.append(False)

len(wdi_clean['Country Code'].unique())

wdi_clean = wdi_clean[country_bool]

len(wdi_clean['Country Code'].unique())

""" Standardize the country codes
"""
country_codes = []
for country in wdi_clean['Country Code'].tolist():
    country_codes.append(pycountry.countries.lookup(country).alpha_3)

wdi_clean.loc[:, 'Country Code'] = country_codes

wdi_clean.describe()

### Saving Data

""" Save the cleaned and standardized data
"""
wdi_clean.to_pickle("../data/interim/WDI.pickle")
pd.DataFrame({
    'Series Code': wdi['Series Code'].unique(),
    'Series Name': wdi['Series Name'].unique()
}).to_pickle("../data/interim/WDI_DD.pickle")

## World Tourism Organization

uwto = pd.read_excel(
    "../data/external/Tourism/UWTO/Arrivals of non resident tourists_visitors, depart.xls"
)

### Grouping by Country

""" 
Create a mask without regions and custom groupings, 
selecting only valid country entries
"""
country_locations = []
row = 0
for country in uwto['COUNTRY']:
    try:
        pycountry.countries.lookup(country)
        country_locations.append(row)
    except LookupError:
        pass
    row += 1
country_locations.append(len(uwto))

""" Create a dataframe for each country code
"""
country_slices = []
for index in range(len(country_locations)):
    try:
        country_slices.append(
            [country_locations[index], country_locations[index + 1] - 1])
    except IndexError:
        pass
nation_frames = []
for s in country_slices:
    nation_frames.append(uwto.loc[s[0]:s[1]])

""" Clean the dataframes and flatten into standard format
"""
clean_nations = []

for nation in nation_frames:
    nation = nation.reset_index()
    nation_name = nation['COUNTRY'][0]

    # list of row indexes indicating a chunk of data (inbound/outbound)
    chunk_headers = []
    row = 0
    for country in nation['COUNTRY']:
        if 'bound tourism' in country:
            chunk_headers.append(row)
        row += 1

    # New dfs on variable groups
    inbound = nation[chunk_headers[0]:chunk_headers[1]]
    outbound = nation[chunk_headers[1]:]

    # rename variables to flatten index
    outbound.loc[1:, 'COUNTRY'] = 'Outbound: ' + outbound[1:]['COUNTRY']
    inbound.loc[1:, 'COUNTRY'] = 'Inbound: ' + inbound[1:]['COUNTRY']

    # combine variable groups and remove index rows
    rejoined = pd.concat([outbound, inbound]).dropna()

    # move indicators to correct column
    rejoined['Indicator'] = rejoined['COUNTRY']

    # set the country name to the correct value
    rejoined['COUNTRY'] = nation_name

    # Fix column order
    rejoined = rejoined[[
        'COUNTRY', 'Indicator', 1995, 1996, 1997, 1998, 1999, 2000, 2001, 2002,
        2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014,
        2015, 2016, 2017
    ]]

    # add to clean list
    clean_nations.append(rejoined)

uwto_clean = pd.concat(clean_nations)

### Clean Values and Group Variables on Means

uwto_clean.head(3)

""" Replace '..' with np.nan for better parsing
"""
uwto_clean = uwto_clean.replace('..', np.NaN)

uwto_clean.head(3)

""" Assign all columns except indicator and country to float dtype
"""
uwto_clean[uwto_clean.columns[2:]] = uwto_clean[uwto_clean.columns[2:]].astype(
    float)

uwto_clean.get_dtype_counts()

""" Select unique
"""
uwto_clean = uwto_clean.groupby(['COUNTRY', 'Indicator']).max()

### Pivot the Data

# Move indicators to the columns index
unstacked = uwto_clean.unstack()
# Name column indexes
unstacked.columns.names = ['Year', 'Indicator']
# move year to row index
restacked = unstacked.stack(level=0, dropna=False)
# move country and year to columns
uwto_pivoted = restacked.reset_index()
# remove indicator label from column index
uwto_pivoted.columns.name = None

uwto_pivoted.head(3)

### Standardize Country Codes

uwto_pivoted['Country Code'] = uwto_pivoted['COUNTRY'].apply(
    lambda x: pycountry.countries.lookup(x).alpha_3)

uwto_pivoted.drop(columns=['COUNTRY'], inplace=True)

### Save the Data

uwto_pivoted.to_pickle("../data/interim/UWTO.pickle")
