==> notebooks/11.ipynb <==
# General Indicators Data Processing (WDI)
## Data Dictionary
| **Code**                     | **Indicator Name**                                                                                                                               |
|--------------------------|----------------------------------------------------------------------------------------------------------------------------------------------|
| SP.DYN.LE00.IN           | Life expectancy at birth, total (years)                                                                                                      |
| EG.ELC.ACCS.ZS           | Access to electricity (% of population)                                                                                                      |
| SH.STA.HYGN.ZS           | People with basic handwashing facilities including soap and water (% of population)                                                          |
| SH.H2O.SMDW.ZS           | People using safely managed drinking water services (% of population)                                                                        |
| SH.H2O.BASW.ZS           | People using at least basic drinking water services (% of population)                                                                        |
| EN.ATM.CO2E.PC           | CO2 emissions (metric tons per capita)                                                                                                       |
| EN.CO2.TRAN.ZS           | CO2 emissions from transport (% of total fuel combustion)                                                                                    |
| ER.LND.PTLD.ZS           | Terrestrial protected areas (% of total land area)                                                                                           |
| DT.DOD.PVLX.GN.ZS        | Present value of external debt (% of GNI)                                                                                                    |
| FB.ATM.TOTL.P5           | Automated teller machines (ATMs) (per 100,000 adults)                                                                                        |
| FB.CBK.BRCH.P5           | Commercial bank branches (per 100,000 adults)                                                                                                |
| FB.CBK.DPTR.P3           | Depositors with commercial banks (per 1,000 adults)                                                                                          |
| FB.CBK.BRWR.P3           | Borrowers from commercial banks (per 1,000 adults)                                                                                           |
| SG.VAW.1549.ZS           | Proportion of women subjected to physical and/or sexual violence in the last 12 months (% of women age 15-49)                                |
| SG.DMK.ALLD.FN.ZS        | Women participating in the three decisions (own health care, major household purchases, and visiting family) (% of women age 15-49)          |
| SG.DMK.SRCR.FN.ZS        | Women making their own informed decisions regarding sexual relations, contraceptive use and reproductive health care  (% of women age 15-49) |
| SH.STA.SUIC.P5           | Suicide mortality rate (per 100,000 population)                                                                                              |
| SH.STA.WASH.P5           | Mortality rate attributed to unsafe water, unsafe sanitation and lack of hygiene (per 100,000 population)                                    |
| SP.DYN.IMRT.IN           | Mortality rate, infant (per 1,000 live births)                                                                                               |
| SH.STA.BRTW.ZS           | Low-birthweight babies (% of births)                                                                                                         |
| SH.ANM.CHLD.ZS           | Prevalence of anemia among children (% of children under 5)                                                                                  |
| SH.STA.ANVC.ZS           | Pregnant women receiving prenatal care (%)                                                                                                   |
| SH.STA.FGMS.ZS           | Female genital mutilation prevalence (%)                                                                                                     |
| SH.PRV.SMOK              | Smoking prevalence, total (ages 15+)                                                                                                         |
| SH.ALC.PCAP.LI           | Total alcohol consumption per capita (liters of pure alcohol, projected estimates, 15+ years of age)                                         |
| IT.NET.BBND.P2           | Fixed broadband subscriptions (per 100 people)                                                                                               |
| IT.NET.USER.ZS           | Individuals using the Internet (% of population)                                                                                             |
| IT.CEL.SETS.P2           | Mobile cellular subscriptions (per 100 people)                                                                                               |
| SM.POP.REFG.OR           | Refugee population by country or territory of origin                                                                                         |
| per_si_allsi.cov_pop_tot | Coverage of social insurance programs (% of population)                                                                                      |
| VC.IHR.PSRC.P5           | Intentional homicides (per 100,000 people)                                                                                                   |
| MS.MIL.TOTL.TF.ZS        | Armed forces personnel (% of total labor force)                                                                                              |
| HD.HCI.OVRL              | Human capital index (HCI) (scale 0-1)                                                                                                        |

import re

import numpy as np
import pandas as pd
import pycountry

%matplotlib inline

pd.set_option('display.float_format', lambda x: '%.3f' % x)

## Load The File

df = pd.read_excel("../data/external/General Indicators/WDI/Data_Extract_From_World_Development_Indicators.xlsx")

df.sample(5)

## Standardize Country Codes

""" Only Select rows with valid country codes
"""
country_locations = []
for country in df['Country Code']:
    try:
        pycountry.countries.lookup(country)
        country_locations.append(True)
    except LookupError:
        country_locations.append(False)
df = df[country_locations]

## Standardize Indexes

df.rename(
    {
        "Time": "Year"
    },
    axis='columns',
    inplace=True)

df.set_index(["Country Code", "Year"], inplace=True)

## Clean Data

### Header

df.drop(["Time Code", "Country Name"],
        axis='columns',
        inplace=True)

c = [ re.search(r"\[(\w+\.)+\w+\]",d)[0].replace("[","").replace("]","") for d in df.columns ]

c_names = {}
for x in range(len(c)):
    c_names[df.columns[x]] = c[x]

df.rename(c_names,axis='columns',inplace=True)

### Data Types

""" Replace '..' with np.nan for better parsing
"""
df = df.replace('..', np.NaN)

df = df.astype(float)

df.sample(5)

## Save Data

df.to_pickle("../data/processed/General_WDI.pickle")
