==> notebooks/12.ipynb <==
# Dataset Joining

import pandas as pd
%matplotlib inline
pd.set_option('display.float_format', lambda x: '%.3f' % x)
pd.set_option('display.max_columns', None)

## Load The Files

Tourism = pd.read_pickle('../data/processed/Tourism.pickle')
Inequality_WIID = pd.read_pickle('../data/processed/Inequality_WIID.pickle')
Inequality_WDI = pd.read_pickle('../data/processed/Inequality_WDI.pickle')
General_WDI = pd.read_pickle('../data/processed/General_WDI.pickle')
Economic_WDI = pd.read_pickle('../data/processed/Economic_WDI.pickle')
Economic_MADDISON = pd.read_pickle('../data/processed/Economic_MADDISON.pickle')
Education_WDI = pd.read_pickle('../data/processed/Education_WDI.pickle')
Labor_WDI = pd.read_pickle('../data/processed/Labor_WDI.pickle')
Urbanization_UNPD = pd.read_pickle('../data/processed/Urbanization_UNPD.pickle')

## Join By Subject

### Inequality

Inequality = Tourism.join(Inequality_WDI).join(Inequality_WIID)

Inequality.sample(5)

### General

General = Tourism.join(General_WDI)

General.sample(5)

### Economic

Economic = Tourism.join(Economic_MADDISON).join(Economic_WDI)

Economic.sample(5)

### Education

Education = Tourism.join(Education_WDI)

Education.sample(5)

### Labor

Labor = Tourism.join(Labor_WDI)

Labor.sample(5)

### Urbanization

Urbanization = Tourism.join(Urbanization_UNPD)

Urbanization.sample(5)

## Full Dataset Join

Full = Tourism.join([
    Inequality_WIID,
    Inequality_WDI,
    General_WDI,
    Economic_MADDISON,
    Education_WDI,
    Labor_WDI,
    Urbanization_UNPD],how="left").join(Economic_WDI.drop(['SI.POV.GINI', 'SI.POV.NAGP'],axis='columns'))

## Save Files

General.to_csv("../data/processed/General.csv")
Labor.to_csv("../data/processed/Labor.csv")
Urbanization.to_csv("../data/processed/Urbanization.csv")
Education.to_csv("../data/processed/Education.csv")
Inequality.to_csv("../data/processed/Inequality.csv")
Economic.to_csv("../data/processed/Economic.csv")
Full.to_csv("../data/processed/Full.csv")
