==> notebooks/16.ipynb <==
# Subject Dataset Cleaning and Imputation

import pandas as pd
from fancyimpute import IterativeImputer
import numpy as np

# Import the datasets
General = pd.read_csv("../data/processed/General.csv")
Labor = pd.read_csv("../data/processed/Labor.csv")
Urbanization = pd.read_csv("../data/processed/Urbanization.csv")
Education = pd.read_csv("../data/processed/Education.csv")
Inequality = pd.read_csv("../data/processed/Inequality.csv")
Economic = pd.read_csv("../data/processed/Economic.csv")

## Remove Sparse Variables

for df in [General,
            Labor,
            Urbanization,
            Education,
            Inequality,
            Economic]:
    ds = df.describe().transpose()
    # columns with <25% values
    sparse = ds[(ds["count"] < 1000)].index.values.tolist()
    print(sparse)
    df.drop(sparse,axis=1,inplace=True)

for df in [General,
            Labor,
            Urbanization,
            Education,
            Inequality,
            Economic]:
    print(len(df.columns))

## Impute Remaining NaNs

for df in [General,
            Labor,
            Urbanization,
            Education,
            Inequality,
            Economic]:
    df.set_index(["Country Code", "Year"], inplace=True)

General_imputed = General.copy()
Labor_imputed = Labor.copy()
Urbanization_imputed = Urbanization.copy()
Education_imputed = Education.copy()
Inequality_imputed = Inequality.copy()
Economic_imputed = Economic.copy()
for df in [General_imputed,
            Labor_imputed,
            Urbanization_imputed,
            Education_imputed,
            Inequality_imputed,
            Economic_imputed]:
    XY_incomplete = df.values
    n_imputations = 5
    XY_completed = []
    for i in range(n_imputations):
        print("Imputing set {} of 5".format(i+1))
        imputer = IterativeImputer(n_iter=10, sample_posterior=True, random_state=i, initial_strategy="median")
        XY_completed.append(imputer.fit_transform(XY_incomplete))

    XY_completed_mean = np.mean(XY_completed, 0)
    XY_completed_std = np.std(XY_completed, 0)
    
    df[:] = XY_completed_mean

## Check The Integrity of Imputed Dataframes

def dfDiff(pair: tuple):
    """
    Takes a tuple of two pandas.DataFrames and returns a dataframe with changed values 
    
    Source: https://wellsr.com/python/pandas-compare-two-data-frames/
    
    IN:
    (
    Index | Col 
    ------|-----
    0     |0  
    1     |1  
    2     |0  
    3     |1  
    4     |1  
    ,
    Index | Col 
    ------|-----
    0     |1  
    1     |1  
    2     |1  
    3     |1  
    4     |0  
    )
    
    OUT:
    Index | Col | Old | New
    ------|-----|-----|-----
    0     |num  |0    |1
    2     |num  |0    |1
    4     |num  |1    |0
    """
    oldFrame, newFrame = pair
    dfBool = (oldFrame != newFrame).stack()
    diff = pd.concat([oldFrame.stack()[dfBool], newFrame.stack()[dfBool]], axis=1)
    diff.columns=["Old", "New"]
    return diff

diffs = [] # This will hold the computed difference dataframes
combinations = [(General_imputed, General),
            (Labor_imputed, Labor),
            (Urbanization_imputed, Urbanization),
            (Education_imputed, Education),
            (Inequality_imputed, Inequality),
            (Economic_imputed, Economic)] # All possible combinations of two dataframes

for pair in combinations:
    diffs.append(dfDiff(pair))

changed = [len(diff) for diff in diffs]

pd.DataFrame(changed).head(20)

## Save the Data Frames

General_imputed.to_csv("../data/production/subject/General.csv")
Labor_imputed.to_csv("../data/production/subject/Labor.csv")
Urbanization_imputed.to_csv("../data/production/subject/Urbanization.csv")
Education_imputed.to_csv("../data/production/subject/Education.csv")
Inequality_imputed.to_csv("../data/production/subject/Inequality.csv")
Economic_imputed.to_csv("../data/production/subject/Economic.csv")
